demo/src/ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.79.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "tenacity>=9.1.2",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "dependency-injector>=4.41.0",
    "ruff>=0.14.4",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
]
